Commands return dyn Error instead of String
This is massively better. How was this not the initial approach? Command
implementors can now use try syntax ('?') on calls to `write!()` which
is much more ergonomic. And allowing any error to be returned is more
flexible, since an application will be using their own error system.
All we use errors for is printing, so there's no change to
functionality.